Complete first section of CSS Project
This commit is related to completing the first section of the CSS
Project. Specifically, this includes the header and the first main
section that's in blue.

Additionally, this commit includes the initial project setup. This
includes the creation of the assets folder and the main HTML and
CSS files.

Lastly, the README file was modified to reflect the overall goals of
the project.